Add payment breakdown (cash/transfer) to order history and decimal quantity support for expenses
Agent-Logs-Url: https://github.com/obongthecreator/120-stand/sessions/21bbd27f-9caf-4aff-a86f-a36f7c54d8c9

+++ includes/class-database.php
@@ -226,7 +226,7 @@ class Stand120_Database {
             expense_date date NOT NULL,
             description varchar(255) NOT NULL,
             amount decimal(10,2) NOT NULL DEFAULT 0,
-            quantity int NOT NULL DEFAULT 1,
+            quantity decimal(10,2) NOT NULL DEFAULT 1,
             total decimal(10,2) NOT NULL DEFAULT 0,
             created_at datetime DEFAULT CURRENT_TIMESTAMP,
             PRIMARY KEY (id),
